+++ WaterFlow3D/Sandbox/Voxel.h
@@ -2,7 +2,7 @@
 #include "Vector.h"
 struct Voxel
 {
-	float Mass = 0;
+	double Mass = 0;
 	Vector Velocity;
 	Vector Force_Internal;
 	Vector Force_External;
